refactor(linter): move word classification into facts
